Move components/git to version-control/git/views
