+++ src/components/editor-footer.tsx
@@ -10,7 +10,7 @@ import { useBufferStore } from "../stores/buffer-store";
 import { useGitStore } from "../stores/git-store";
 import { useUIState } from "../stores/ui-state-store";
 import { getGitStatus } from "../utils/git";
-import GitBranchManager from "./git/git-branch-manager";
+import GitBranchManager from "../version-control/git/views/git-branch-manager";
 
 const EditorFooter = () => {
   const buffers = useBufferStore.use.buffers();
