+++ src/version-control/git/views/git-stash-manager.tsx
